tools/printer-service/requirements.txt
appdirs==1.4.4
argcomplete==3.6.3
importlib_resources==6.5.2
pillow==11.3.0
python-barcode==0.16.1
python-escpos==3.1
pyusb==1.3.1
PyYAML==6.0.3
qrcode==8.2
six==1.17.0
zipp==3.23.0
flask>=3.0.0